--- Step Tracker/DashBoardView.swift
@@ -1,179 +1,0 @@
-//
-//  DashBoardView.swift
-//  Step Tracker
-//
-//  Created by Jeremy Cinq-Mars on 2024-04-29.
-//
-
-import SwiftUI
-import Charts
-
-enum HealthMetricContext: CaseIterable, Identifiable {
-    // could add another case if you want to track something else like calories, or sleep
-    case steps, weight
-    var id: Self { self }
-    
-    var title: String {
-        switch self {
-        case.steps:
-            return "Steps"
-        case.weight:
-            return "Weight"
-        }
-    }
-}
-
-
-struct DashBoardView: View {
-    
-    @Environment(HealthKitManager.self) private var hkManager
-    @AppStorage("hasSeenPermissionPriming") private var hasSeenPermissionPriming = false
-    @State private var isShowingPermissionPrimingSheet = false
-    @State private var selectedStat: HealthMetricContext = .steps
-    @State private var rawSelectedDate: Date?
-    var isSteps: Bool {selectedStat == .steps}
-    
-    var avgStepCount: Double {
-        guard !hkManager.stepData.isEmpty else { return 0 }
-        let totalSteps = hkManager.stepData.reduce(0) { $0 + $1.value}
-        return totalSteps/Double(hkManager.stepData.count)
-    }
-    
-    var selectedHealthMetric: HealthMetric? {
-        guard let rawSelectedDate else { return nil }
-        return hkManager.stepData.first {
-            Calendar.current.isDate(rawSelectedDate, inSameDayAs: $0.date)
-        }
-    }
-    
-    var body: some View {
-        NavigationStack {
-            ScrollView {
-                VStack(spacing: 20) {
-                    Picker("Selected Stat", selection: $selectedStat) {
-                        ForEach(HealthMetricContext.allCases) {
-                            Text($0.title)
-                        }
-                    }
-                    .pickerStyle(.segmented)
-                    VStack {
-                        NavigationLink(value: selectedStat){
-                            HStack {
-                                VStack(alignment: .leading) {
-                                    Label("Steps", systemImage: "figure.walk")
-                                        .font(.title3.bold())
-                                        .foregroundStyle(.pink)
-                                    
-                                    Text("Avg: \(Int(avgStepCount)) Steps")
-                                        .font(.caption)
-                                }
-                                
-                                Spacer()
-                                
-                                Image(systemName: "chevron.right")
-                            }
-                        }
-                        .foregroundStyle(.secondary)
-                        .padding(.bottom, 12)
-                        
-                        Chart {
-                            if let selectedHealthMetric {
-                                RuleMark(x: .value("Selected Metric", selectedHealthMetric.date, unit: .day))
-                                    .foregroundStyle(Color.secondary.opacity(0.3))
-                                    .offset(y: -10)
-                                    .annotation(position: .top,
-                                                spacing: 0,
-                                                overflowResolution: .init(x: .fit(to: .chart), y: .disabled)) { annotationView }
-                            }
-                            RuleMark(y: .value("Avegare", avgStepCount))
-                                .foregroundStyle(Color.secondary)
-                                .lineStyle(.init(lineWidth: 1, dash: [5]))
-                            
-                            ForEach(hkManager.stepData) { steps in
-                                BarMark(x: .value("Date", steps.date, unit: .day),
-                                        y: .value("Steps", steps.value)
-                                )
-                                .foregroundStyle(Color.pink.gradient)
-                                .opacity(rawSelectedDate == nil || steps.date == selectedHealthMetric?.date ? 1.0 : 0.3)
-                            }
-                        }
-                        .frame(height: 150)
-                        .chartXSelection(value: $rawSelectedDate.animation(.easeInOut))
-                        .chartXAxis {
-                            AxisMarks {
-                                AxisValueLabel(format: .dateTime.month(.defaultDigits).day())
-                            }
-                        }
-                        .chartYAxis {
-                            AxisMarks { value in
-                                AxisGridLine()
-                                    .foregroundStyle(Color.secondary.opacity(0.3))
-                                AxisValueLabel((value.as(Double.self) ?? 0).formatted(.number.notation(.compactName)))
-                            }
-                        }
-                    }
-                    .padding()
-                    .background(RoundedRectangle(cornerRadius: 12).fill(Color(.secondarySystemBackground)))
-                    
-                    VStack(alignment: .leading) {
-                        VStack(alignment: .leading) {
-                            Label("Averages", systemImage: "calendar")
-                                .font(.title3.bold())
-                                .foregroundStyle(.pink)
-                            
-                            Text("Last 28 Days")
-                                .font(.caption)
-                                .foregroundStyle(.secondary)
-                        }
-                        .padding(.bottom, 12)
-                        
-                        RoundedRectangle(cornerRadius: 12)
-                            .foregroundStyle(.secondary)
-                            .frame(height: 240)
-                    }
-                    .padding()
-                    .background(RoundedRectangle(cornerRadius: 12).fill(Color(.secondarySystemBackground)))
-                }
-            }
-            .padding()
-            .task {
-//                await hkManager.addSimulatorData()
-                await hkManager.fetchStepCount()
-                isShowingPermissionPrimingSheet = !hasSeenPermissionPriming
-            }
-            .navigationTitle("Dashboard")
-            .navigationDestination(for: HealthMetricContext.self) { metric in
-                HealthDataListView(metric: metric)
-            }
-            .sheet(isPresented: $isShowingPermissionPrimingSheet, onDismiss: {
-                //fetch health data
-            }, content: {
-                HKPermissionPrimmingView(hasSeen: $hasSeenPermissionPriming)
-            })
-        }
-        .tint(isSteps ? .pink : .indigo)
-    }
-    
-    var annotationView: some View {
-        VStack(alignment: .leading) {
-            Text(selectedHealthMetric?.date ?? .now, format: 
-                    .dateTime.weekday(.abbreviated).month(.abbreviated).day())
-                .font(.footnote.bold())
-                .foregroundStyle(.secondary)
-            
-            Text(selectedHealthMetric?.value ?? 0, format: .number.precision(.fractionLength(0)))
-                .fontWeight(.heavy)
-                .foregroundStyle(.pink)
-        }
-        .padding(12)
-        .background(RoundedRectangle(cornerRadius: 4)
-            .fill(Color(.secondarySystemBackground))
-            .shadow(color: .secondary.opacity(0.3), radius: 2, x: 2, y: 2)
-        )
-    }
-}
-
-#Preview {
-    DashBoardView()
-        .environment(HealthKitManager())
-}
