+++ Step Tracker/Manager/HealthKitManager.swift
